final shortenLink area as new component
