+++ client/src/components/output/Output.jsx
@@ -1,0 +1,25 @@
+import React, { useEffect, useState } from 'react'
+import { Icon } from '@iconify/react';
+import './output.css'
+import { inputUrl, shrinkIt, disableInputArea } from '../../handler/inputTextHandler.js'
+
+const Output = ({setIsCustom}) => {
+    const CopyToClipboard = async() => {
+        const text = document.getElementById("output").value
+        await navigator.clipboard.writeText(text)
+        window.alert("copied to clipboard")
+    }
+  return (
+    <>
+    <div className='output-area' id='outputArea'>
+        <input className='url-type' id='output' name='shortUrl' type='text' disabled></input>
+        {/* input field is disabled by default */}
+        <button id='btn-copy' className='btn-black' onClick={() => CopyToClipboard()} title='copy'><Icon icon="material-symbols:content-copy-outline" width="1.3rem" height="1.3rem" inline={true} /></button>
+        <button className='btn-black' id='btn-customise' onClick={() => setIsCustom(true)} title='customise'><Icon icon="system-uicons:write" width="1.3rem" height="1.3rem" inline={true} /></button>
+        <button id='btn-shorten2' className='btn-black' onClick={() => disableInputArea(false)}>shorten another url</button>
+    </div>
+    </>
+  )
+}
+
+export default Output
