+++ client/src/components/output/output.css
@@ -1,0 +1,21 @@
+.output-area {
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    place-items: flex-start;
+    gap: 0.2rem;
+    display: none;
+}
+
+.output-area input {
+    width: 31%;
+    height: max-content;
+}
+
+#btn-copy, #btn-customise {
+    width: 5%;
+}
+
+#btn-shorten2 {
+    width: auto;
+}
