+++ client/src/components/shortener/shortener.css
@@ -101,25 +101,3 @@
     left: 0;
     padding: 0rem;
 }
-
-.output-area {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    place-items: flex-start;
-    gap: 0.2rem;
-    display: none;
-}
-
-.output-area input {
-    width: 31%;
-    height: max-content;
-}
-
-#btn-copy, #btn-customise {
-    width: 5%;
-}
-
-#btn-shorten2 {
-    width: auto;
-}
