+++ client/src/handler/inputTextHandler.js
@@ -16,15 +16,15 @@ export const shrinkIt = (useStateFunc, idVal) => {
 
 export const disableInputArea = (val) => {
   if(val){
-    document.getElementById("outputArea").style.display = 'flex'
-    document.getElementById("btn_shorten").style.display = 'none'
-    document.getElementById("url-type").setAttribute('disabled','disabled');
+    window.document.getElementById("outputArea").style.display = 'flex'
+    window.document.getElementById("btn_shorten").style.display = 'none'
+    window.document.getElementById("url-type").setAttribute('disabled','disabled');
   }
   else {
-    document.getElementById("outputArea").style.display = 'none'
-    document.getElementById("btn_shorten").style.display = 'inline-block'
-    document.getElementById("url-type").removeAttribute('disabled')
-    document.getElementById("url-type").value=''
+    window.document.getElementById("outputArea").style.display = 'none'
+    window.document.getElementById("btn_shorten").style.display = 'inline-block'
+    window.document.getElementById("url-type").removeAttribute('disabled')
+    window.document.getElementById("url-type").value=''
   }
 }
